feat(safety): give Emergency Contacts its own URL; reorder nav
Move the Emergency Contacts view off the bare /safety index onto
/safety/emergency-contacts so every Safety subtab has its own URL,
matching the membership-ops/program-ops tab pattern. /safety now
redirects to the first tab. Drop the redundant page title that just
echoed the tab name.

Also move "Safety" in the left nav to sit directly under "My Household".
